Introduce PicrossSegment and BlockGroup classes
PicrossSegments represent the rows and columns of a picross puzzle.
BlockGroups represent the numbers next to rows and above columns
indicating how many blocks are to be filled in.

PicrossSegments have detailed initialization-based methods and basic
methods and properties for accessing class data. BlockGroup currently
only includes a constructor and a few get/set properties.